Working sending transactions to the blockchain, need to sync with Vali for next step.
Currently the sending of the transactions to the blockchain is working, so are the login functionalities.

I have tested some of the blockchain operations and they seem to propagate to the blockchain but we aren't seeing the result we need, might be from the SM itself, I need to sync with Vali.